More user -> account changes

+++ MSAL/src/MSALErrorConverter.m
@@ -101,7 +101,7 @@ static NSDictionary *s_userInfoKeyMapping;
         }
         else
         {
-            MSID_LOG_ERROR(nil, @"MSALErrorConverter could not find the error code mapping entry for domain (%@) + error code (%ld).", msidError.domain, msidError.code);
+            MSID_LOG_ERROR(nil, @"MSALErrorConverter could not find the error code mapping entry for domain (%@) + error code (%ld).", msidError.domain, (long)msidError.code);
         }
     }
